+++ examples/medplum-provider/src/App.tsx
@@ -192,7 +192,9 @@ export function App(): JSX.Element | null {
                     : []),
                   { icon: <IconUserPlus />, label: 'New Patient', href: '/onboarding' },
                   { icon: <IconApps />, label: 'Integrations', href: '/integrations' },
-                  ...(hasBilling ? [{ icon: <IconReceipt2 />, label: 'Billing Setup', href: '/settings/billing' }] : []),
+                  ...(hasBilling
+                    ? [{ icon: <IconReceipt2 />, label: 'Billing Settings', href: '/Settings/Billing' }]
+                    : []),
                   ...(hasDoseSpot
                     ? [
                         {
@@ -303,7 +305,7 @@ export function App(): JSX.Element | null {
               {hasScriptSure && <Route path="/scriptsure" element={<ScriptSurePage />} />}
               <Route path="/integrations" element={<IntegrationsPage />} />
               {/* Must precede the /:resourceType catch-alls below */}
-              {hasBilling && <Route path="/settings/billing" element={<BillingSetupPage />} />}
+              {hasBilling && <Route path="/Settings/Billing/:tab?" element={<BillingSetupPage />} />}
               <Route path="/smart-health-link" element={<SmartHealthLinkImportPage />} />
               <Route path="/:resourceType" element={<SearchPage />} />
               <Route path="/:resourceType/new" element={<ResourceCreatePage />} />
